requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.2.0
requests==2.19.1
urllib3==1.23
